refactor add_css_to_component to not save customStyleTag in component.$$
